requirements.txt
python-dotenv
mysql-connector-python
openpyxl
python-docx 
mcp
requests